Try out static-values option in all phases
As a temporary way of trying out this option, we change the diffblue.yml
for Dragonfly directly instead of using the update-diffblue-yml script
from dashboard. The next time the script is run, this change will be
overwritten, so we will have to come up with something more robust by
then. Eventually we will want to use this option on all benchmarks.